+++ a3fe/run/system_prep.py
@@ -17,7 +17,6 @@ from typing import Optional as _Optional
 import BioSimSpace.Sandpit.Exscientia as _BSS
 from pydantic import BaseModel as _BaseModel
 from pydantic import Field as _Field
-from pydantic import field_validator as _field_validator
 
 from ..read._process_bss_systems import rename_lig as _rename_lig
 from ._utils import check_has_wat_and_box as _check_has_wat_and_box
@@ -36,10 +35,6 @@ class SystemPreparationConfig(_BaseModel):
         Whether to use SLURM for the preparation.
     forcefields : dict
         Forcefields to use for the ligand, protein, and water.
-    lig_net_charge : int
-        Net charge of the ligand. It is not recommended to use A3FE for charged ligands
-        as it uses reaction field electrostatics and does not keep the box netural or apply
-        corrections. Run at your own risk!
     water_model : str
         Water model to use.
     ion_conc : float
@@ -77,7 +72,6 @@ class SystemPreparationConfig(_BaseModel):
         "protein": "ff14SB",
         "water": "tip3p",
     }
-    lig_net_charge: int = _Field(0, ge=-5, le=5)
     water_model: str = "tip3p"
     ion_conc: float = _Field(0.15, ge=0, lt=1)  # M
     steps: int = _Field(1000, gt=0, lt=100_000)  # This is the default for _BSS
@@ -181,18 +175,6 @@ class SystemPreparationConfig(_BaseModel):
             ],
         },
     }
-
-    @_field_validator("lig_net_charge")
-    def validate_lig_net_charge(lig_net_charge: int) -> int:
-        """Validator for the ligand net charge."""
-        if lig_net_charge != 0:
-            # Raise a user warning if the ligand is charged
-            _warnings.warn(
-                "Warning: A3FE uses reaction field electrostatics and does not keep the box neutral or apply corrections. "
-                "This can lead to artefacts in the results. Run charged ligands at your own risk!"
-            )
-
-        return lig_net_charge
 
     class Config:
         """
@@ -316,11 +298,21 @@ def parameterise_input(
     lig_sys = _BSS.IO.readMolecules(f"{input_dir}/ligand.sdf")
     # Ensure that the ligand is named "LIG"
     _rename_lig(lig_sys, "LIG")
-    param_lig = _BSS.Parameters.parameterise(
-        molecule=lig_sys[0],
-        forcefield=cfg.forcefields["ligand"],
-        net_charge=cfg.lig_net_charge,
-    ).getMolecule()
+    # Check charge of the ligand
+    lig = lig_sys[0]
+    lig_charge = round(lig.charge().value())
+    if lig_charge != 0:
+        _warnings.warn(
+            f"Ligand has a charge of {lig_charge}. Co-alchemical ion approach will be used."
+            " Ensure that your box is large enough to avoid artefacts."
+        )
+
+    # Only include ligand charge if we're using gaff (OpenFF doesn't need it)
+    param_args = {"molecule": lig, "forcefield": cfg.forcefields["ligand"]}
+    if "gaff" in cfg.forcefields["ligand"]:
+        param_args["net_charge"] = lig_charge
+
+    param_lig = _BSS.Parameters.parameterise(**param_args).getMolecule()
 
     # If bound, then parameterise the protein and waters and add to the system
     if leg_type == _LegType.BOUND: