+++ docs/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Change Log
 ===============
 
+0.2.0
+====================
+
+- Added ability to run charged compounds using the co-alchemical ion approach. Ligand net charge is detected from the input files and the config is checked to ensure that PME is used for electrostatics.
+- Added parameterisation tests.
+- Fixed bug arising from incorrect removal of "is_equilibrated" method (37de921)
+- Made detection of username more robust (see https://github.com/michellab/a3fe/issues/8)
+
 0.1.3
 ====================